+++ node/globals.ts
@@ -76,7 +76,6 @@ declare global {
     next: Maybe<string>
     report: Record<string, number>
     count: number
-    disableDraftRoutes: boolean
     disableStringRoutes: string
   }