+++ node/middlewares/generateMiddlewares/generateRewriterRoutes.ts
@@ -21,7 +21,7 @@ const LIST_LIMIT = 300
 
 type RoutesByBinding = Record<string, Record<string, Route[]>>
 
-const createRoutesByBinding = (routes: Internal[], report: Record<string, number>, storeBindings: Binding[], disableDraftRoutes: boolean, disableStringRoutes: string) => {
+const createRoutesByBinding = (routes: Internal[], report: Record<string, number>, storeBindings: Binding[], disableStringRoutes: string) => {
   const storeBindingsIds = storeBindings.map(({ id }) => id)
   return routes.reduce(
     (acc, internal) => {
@@ -31,9 +31,6 @@ const createRoutesByBinding = (routes: Internal[], report: Record<string, number
         internal.type !== 'product' &&
         !internal.disableSitemapEntry &&
         storeBindingsIds.includes(internal.binding)
-      if (disableDraftRoutes) {
-        validRoute = validRoute && !internal.from.startsWith('/_draft')
-      }
       if (disableStringRoutes !== '') {
         validRoute = validRoute && !internal.from.includes('/' + disableStringRoutes)
       }
@@ -101,7 +98,6 @@ export async function generateRewriterRoutes(ctx: EventContext, nextMiddleware: 
   const { clients: { rewriter, tenant } , body } = ctx
   const {
     count,
-    disableDraftRoutes,
     disableStringRoutes,
     generationId,
     next,
@@ -113,7 +109,7 @@ export async function generateRewriterRoutes(ctx: EventContext, nextMiddleware: 
   const responseNext = response.next
 
   const storeBindings = await getStoreBindings(tenant)
-  const routesByBinding = createRoutesByBinding(routes, report, storeBindings, disableDraftRoutes, disableStringRoutes)
+  const routesByBinding = createRoutesByBinding(routes, report, storeBindings, disableStringRoutes)
 
   await Promise.all(
     Object.keys(routesByBinding).map(saveRoutes(routesByBinding, count, ctx.clients))
@@ -122,7 +118,6 @@ export async function generateRewriterRoutes(ctx: EventContext, nextMiddleware: 
   if (responseNext) {
     const payload: RewriterRoutesGenerationEvent = {
       count: count + 1,
-      disableDraftRoutes,
       disableStringRoutes,
       generationId,
       next: responseNext,