+++ node/middlewares/generateMiddlewares/generateSitemap.test.ts
@@ -66,7 +66,6 @@ describe('Test generate sitemap', () => {
       state: {
         ...state.object,
         settings: {
-          disableDraftRoutes: false,
           disableStringRoutes: '',
           enableAppsRoutes: true,
           enableNavigationRoutes: true,
@@ -89,7 +88,6 @@ describe('Test generate sitemap', () => {
 
   it('Should send only enabled events', async () => {
     context.state.settings = {
-      disableDraftRoutes: false,
       disableStringRoutes: '',
       enableAppsRoutes: true,
       enableNavigationRoutes: true,
@@ -103,7 +101,6 @@ describe('Test generate sitemap', () => {
 
     jest.clearAllMocks()
     context.state.settings = {
-      disableDraftRoutes: false,
       disableStringRoutes:'',
       enableAppsRoutes: false,
       enableNavigationRoutes: false,