+++ node/middlewares/generateMiddlewares/generateSitemap.ts
@@ -29,12 +29,10 @@ const DEFAULT_REWRITER_ROUTES_PAYLOAD = {
 
 export async function generateSitemap(ctx: EventContext) {
   const { clients: { events }, body: { generationId }, state: { settings } }  = ctx
-  const disableDraftRoutes = settings.disableDraftRoutes
   const disableStringRoutes = settings.disableStringRoutes
   if (settings.enableNavigationRoutes) {
     events.sendEvent('', GENERATE_REWRITER_ROUTES_EVENT, {
       ...DEFAULT_REWRITER_ROUTES_PAYLOAD,
-      disableDraftRoutes,
       disableStringRoutes,
       generationId,
     } as RewriterRoutesGenerationEvent)